requirements.txt
requests
beautifulsoup4
langchain
langchain-community
langchain-openai
chromadb
tiktoken
gradio
python-dotenv
